pyproject.toml
[project]
name = "moshi-core"
version = "23.8.28"
description = "Moshi is a spoken language tutor."
readme = "README.md"
requires-python = ">=3.7"
license = {text = "Proprietary"}
keywords = ["language", "ai", "moshi"]
authors = [
  {name = "Eric Kalosa-Kenyon", email = "moshi.feedback@gmail.com"}
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python :: 3",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "Topic :: Education",
]
dependencies = [
  "av",
  "firebase_admin",
  "google-auth",
  "google-cloud-firestore",
  "google-cloud-logging",
  "google-cloud-translate",
  "google-cloud-texttospeech",
  "google-cloud-secret-manager",
  "loguru",
  "numpy",
  "openai",
]

[project.optional-dependencies]
tests = [
  "httpx",
  "pytest",
  "pytest-cov",
]
dev = [
  "black",
  "flake8",
  "isort",
  "pyfiglet"
]

[project.urls]
Homepage = "https://www.chatmoshi.com"
Repository = "https://github.com/ekalosak/moshi-core.git"


[tool.pytest.ini_options]
markers = [
  "slow: runs more than a second or two to run",
  "oai: OpenAI APIs ($)",
  "gcp: Google Cloud APIs ($)",
  "aio: uses asyncio, async and await",
]
# addopts = [
#   "--cov=moshi"
# ]
